% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIdata.R
\docType{data}
\name{PIdata1}
\alias{PIdata1}
\title{PIdata includes two datasets, PIdata1 and PIdata2, which are simulation data.}
\format{Objects of class data.frame, and missing values are coded as -999.}
\usage{
data(PIdata)
}
\description{
PIdata1 and PIdata2 are from simple random samples and stratified simple random samples, respectively. The two variables of "samp.weight" and "strata" should be included on the dataset of stratified random samples. The variable of "strata.frac" is needed for the design-based variance calculation with the option "sample.design=2".
}
\examples{
data(PIdata)
head(PIdata1)
head(PIdata2)
}
\keyword{datasets}
